Move PrivateRoute and PublicRoute from components
